+++ client/proxgui.h
@@ -41,7 +41,7 @@ extern bool GridLocked;
 //Operations defined in data_operations
 //int autoCorr(const int* in, int *out, size_t len, int window);
 int AskEdgeDetect(const int *in, int *out, int len, int threshold);
-int AutoCorrelate(const int *in, int *out, size_t len, int window, bool SaveGrph, bool verbose);
+int AutoCorrelate(const int *in, int *out, size_t len, size_t window, bool SaveGrph, bool verbose);
 int directionalThreshold(const int *in, int *out, size_t len, int8_t up, int8_t down);
 void save_restoreGB(uint8_t saveOpt);
